add scroll bar
