+++ demos/0-interacion-pinch-pan-xy.html
@@ -41,6 +41,19 @@ $.getJSON('./data/cereal.json', function(data) {
       console.log('pinch', this.xRange, this.yRange);
     }
   });
+  chart.scrollBar({
+    mode: 'xy',
+    xStyle: {
+      backgroundColor: '#e8e8e8',
+      fillerColor: '#808080',
+      offsetY: -2
+    },
+    yStyle: {
+      backgroundColor: '#e8e8e8',
+      fillerColor: '#808080',
+      offsetX: 2
+    }
+  });
   chart.render();
 });
 </script>
